don't output curl's progress (#7)
Make curl silent using the --silent option. This was useful for letting the user know that the github command didn't simply hang or get stuck, but the output was dirtying the console. As a substitute, simply output that the command is now querying GitHub and may take a while. Also, remove an extraneous echo that was accidentally left in after debugging.